requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
pillow>=9.5.0
pandas>=2.0.0